+++ src/main/java/traben/entity_texture_features/mixin/client/entity/blockEntity/MixinChestBlockEntityRenderer.java
@@ -5,7 +5,6 @@ import net.minecraft.block.Block;
 import net.minecraft.block.BlockState;
 import net.minecraft.block.DoubleBlockProperties;
 import net.minecraft.block.entity.BlockEntity;
-import net.minecraft.block.entity.ChestBlockEntity;
 import net.minecraft.block.enums.ChestType;
 import net.minecraft.client.MinecraftClient;
 import net.minecraft.client.block.ChestAnimationProgress;
@@ -22,9 +21,7 @@ import net.minecraft.entity.decoration.ArmorStandEntity;
 import net.minecraft.util.Identifier;
 import net.minecraft.util.Nameable;
 import net.minecraft.world.World;
-import org.spongepowered.asm.mixin.Final;
 import org.spongepowered.asm.mixin.Mixin;
-import org.spongepowered.asm.mixin.Shadow;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.ModifyArg;
@@ -38,37 +35,6 @@ import static traben.entity_texture_features.client.ETFClient.ETFConfigData;
 
 @Mixin(ChestBlockEntityRenderer.class)
 public abstract class MixinChestBlockEntityRenderer<T extends BlockEntity & ChestAnimationProgress> implements BlockEntityRenderer<T> {
-
-    @Final
-    @Shadow
-    private ModelPart doubleChestLeftLid;
-    @Final
-    @Shadow
-    private ModelPart doubleChestLeftLatch;
-    @Final
-    @Shadow
-    private ModelPart doubleChestLeftBase;
-
-    @Final
-    @Shadow
-    private ModelPart doubleChestRightLid;
-    @Final
-    @Shadow
-    private ModelPart doubleChestRightLatch;
-    @Final
-    @Shadow
-    private ModelPart doubleChestRightBase;
-
-    @Final
-    @Shadow
-    private ModelPart singleChestLid;
-    @Final
-    @Shadow
-    private ModelPart singleChestLatch;
-    @Final
-    @Shadow
-    private ModelPart singleChestBase;
-
 
     @ModifyArg(method = "render(Lnet/minecraft/block/entity/BlockEntity;FLnet/minecraft/client/util/math/MatrixStack;Lnet/minecraft/client/render/VertexConsumerProvider;II)V",
             at = @At(value = "INVOKE", target = "Lnet/minecraft/client/render/block/entity/ChestBlockEntityRenderer;render(Lnet/minecraft/client/util/math/MatrixStack;Lnet/minecraft/client/render/VertexConsumer;Lnet/minecraft/client/model/ModelPart;Lnet/minecraft/client/model/ModelPart;Lnet/minecraft/client/model/ModelPart;FII)V"),
@@ -87,7 +53,7 @@ public abstract class MixinChestBlockEntityRenderer<T extends BlockEntity & Ches
     @Inject(method = "render(Lnet/minecraft/block/entity/BlockEntity;FLnet/minecraft/client/util/math/MatrixStack;Lnet/minecraft/client/render/VertexConsumerProvider;II)V",
             at = @At(value = "INVOKE", target = "Lnet/minecraft/client/util/SpriteIdentifier;getVertexConsumer(Lnet/minecraft/client/render/VertexConsumerProvider;Ljava/util/function/Function;)Lnet/minecraft/client/render/VertexConsumer;",
                     shift = At.Shift.AFTER), locals = LocalCapture.CAPTURE_FAILSOFT)
-    private void etf$getChestTexture(T entity, float tickDelta, MatrixStack matrices, VertexConsumerProvider vertexConsumers, int light, int overlay, CallbackInfo ci, World world, boolean bl, BlockState blockState, ChestType chestType, Block block, AbstractChestBlock abstractChestBlock, boolean bl2, float f, DoubleBlockProperties.PropertySource propertySource, float g, int i, SpriteIdentifier spriteIdentifier) {
+    private void etf$getChestTexture(T entity, float tickDelta, MatrixStack matrices, VertexConsumerProvider vertexConsumers, int light, int overlay, CallbackInfo ci, World world, boolean bl, BlockState blockState, ChestType chestType, Block block, AbstractChestBlock<?> abstractChestBlock, boolean bl2, float f, DoubleBlockProperties.PropertySource<?> propertySource, float g, int i, SpriteIdentifier spriteIdentifier) {
 
         //hopefully works in modded scenarios, assumes the mod dev uses the actual vanilla code process and texture pathing rules
         String nameSpace = spriteIdentifier.getTextureId().getNamespace();
@@ -97,15 +63,16 @@ public abstract class MixinChestBlockEntityRenderer<T extends BlockEntity & Ches
         if (ETFConfigData.enableCustomTextures) {
             etf$chestStandInDummy = new ArmorStandEntity(EntityType.ARMOR_STAND, MinecraftClient.getInstance().world);
             etf$chestStandInDummy.setPos(entity.getPos().getX(), entity.getPos().getY(), entity.getPos().getZ());
-            String identifier = entity.getPos().toString();
+            String identifier = "chest" + entity.getPos().toString() + chestType.asString();
             if (entity instanceof Nameable nameable) {
                 etf$chestStandInDummy.setCustomName(nameable.getCustomName());
                 etf$chestStandInDummy.setCustomNameVisible(nameable.hasCustomName());
                 if(nameable.hasCustomName()) {
+                    //noinspection ConstantConditions
                     identifier += nameable.getCustomName().asString();
                 }
             }
-            //chests don't have uuid so set UUID from something repeatable I chose from block pos
+            //chests don't have uuid so set UUID from something repeatable this uses blockPos chestType & container name
             etf$chestStandInDummy.setUuid(UUID.nameUUIDFromBytes(identifier.getBytes()));
         }
     }